Add script to automate updating common/ in our projects via a PR
You can run the update-common-everywhere.sh as follows:

    ./update-common-everywhere.sh -u mbaldessari -r multicloud-gitops -p

The above will clone hybrid-cloud-patterns/{common,multicloud-gitops},
do the necessary subtree merging magic. Push the change to a branch in
your github user forked repo (repo needs to be forked in advance) and
create a PR with the changes.
It will bail out if the diff between common proper and common subtree
is non-empty.

It requires the 'gh' CLI utility to be installed.